Implemented xml parsing functionality (using SAX).
Finished implementing xml reading functionality, but code needs to be
reorganized so that the Workout and Exercise classes are shared between
the XML parser and the rest of the program. I didn't completely
understand how I wanted to structure this part of the program when I was
writing the code, so I made some unnecessary redundancies.