Fixed previous bugs and made a lot of additional progress.
Made a new class "Main" as a new point of execution for the program.
This allows me to create an instance of the Drill_Sergeant class within
it.  Before it was a static class since it contains the GUI components,
but that was no longer working out.

Made a lot of progress on Active Workout screen.  It is almost fully
functional and can currently run through an entire workout without
issue.