Fixed bug in Preview window.
Fixed Bug:  Preview window's table of exercises was not being reset
before user left the "Edit Workout" screen.

Also, tried to finish implementing "Delete" functionality, but the
Delete button on Preview Window is only effective at deleting the
exercises from the table in the Preview window.   Delete button may need
to be relocated to the main application window to facilitate deletion of
exercises in the actual Workout object.